requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
scipy>=1.11
statsmodels>=0.14
imbalanced-learn>=0.12
# optional (skip if not available)
xgboost>=2.0
lightgbm>=4.3
shap>=0.45.0
matplotlib>=3.8
plotly>=5.22
streamlit>=1.36
pyyaml>=6.0.1
joblib>=1.4
seaborn>=0.13.2